ragx-gemini/pyproject.toml
[project]
name = "ragx-gemini"
version = "0.1.0"
description = "Basic RAG app using Google AI"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "google-genai (>=1.12.1,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "langchain[google-genai] (>=0.3.24,<0.4.0)",
    "langchain-google-genai (>=2.1.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
